Created several components in App.js in separate files
