+++ src/App.js
@@ -1,9 +1,7 @@
-/** @format */
-
 import D3DirectedGraph from "./D3DirectedGraph";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import { Link } from "react-router-dom";
 import "bulma/css/bulma.css";
+import Header from "./components/Header";
 
 function App() {
   return (
@@ -11,7 +9,6 @@ function App() {
       <div>
         <Header />
         <Main />
-        <Footer />
       </div>
     </Router>
   );
@@ -19,19 +16,16 @@ function App() {
 
 function Main() {
   return (
-    <section
-      className="section has-background-white-bis"
-      style={{ minHeight: "1000px" }}
-    >
+    <section>
       <Switch>
         <Route path="/" exact>
-          <div>
+          <div className="section has-background-white-bis">
             <div className="columns is-centered">
               <AreaTab />
               <Search />
             </div>
-            <D3DirectedGraph />
           </div>
+          <D3DirectedGraph />
         </Route>
         <Route path="/usage">
           <div>使い方だよ！</div>
@@ -44,25 +38,6 @@ function Main() {
         </Route>
       </Switch>
     </section>
-  );
-}
-
-function Header() {
-  return (
-    <div className="hero is-small is-primary">
-      <section className="section hero-body">
-        <div className="container">
-          <div className="columns">
-            <div className="column is-one-third is-offset-one-third">
-              <h1 className="title is-3 has-text-centered">
-                2021QiitaViz-legends
-              </h1>
-            </div>
-            <Navigation />
-          </div>
-        </div>
-      </section>
-    </div>
   );
 }
 
@@ -119,52 +94,6 @@ function Search() {
         </div>
       </div>
     </div>
-  );
-}
-function Navigation() {
-  return (
-    <div className="column">
-      <div className={`dropdown is-hoverable`}>
-        <div className="dropdown-trigger">
-          <button
-            className="button is-primary"
-            aria-haspopup="true"
-            aria-controls="dropdown-menu"
-          >
-            <span className="has-text-weight-bold">メニュー</span>
-            <span className="icon is-small">
-              <i className="fas fa-angle-down"></i>
-            </span>
-          </button>
-        </div>
-        <div className="dropdown-menu" id="dropdown-menu" role="menu">
-          <div className="dropdown-content">
-            <Link className="dropdown-item" to="/">
-              ホーム
-            </Link>
-            <Link className="dropdown-item" to="/usage">
-              使い方
-            </Link>
-            <Link className="dropdown-item" to="/description">
-              このサイトでできること
-            </Link>
-            <Link className="dropdown-item" to="/references">
-              参考文献
-            </Link>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function Footer() {
-  return (
-    <footer className="footer has-background-white-ter">
-      <div className="content has-text-centered">
-        <p>2021 QiitaViz</p>
-      </div>
-    </footer>
   );
 }
