+++ src/components/Footer.js
@@ -1,0 +1,13 @@
+import "bulma/css/bulma.css";
+
+const Footer = () => {
+  return (
+    <footer className="footer has-background-white-ter">
+      <div className="content has-text-centered">
+        <p>2021 QiitaViz</p>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;
